create HomeScreen file
